Separated lobster_doxygen.py into different modules.
